+++ ta4j-core/src/main/java/org/ta4j/core/TimeSeries.java
@@ -25,6 +25,7 @@ package org.ta4j.core;
 import org.ta4j.core.num.Num;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.Duration;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
@@ -144,27 +145,133 @@ public interface TimeSeries extends Serializable {
      * Exceeding bars are removed.
      * @param bar the bar to be added
      * @see TimeSeries#setMaximumBarCount(int)
-     * @deprecated use the {@link #addBar(Duration, ZonedDateTime)}  other addBar functions} to add and convert data directly
+     * @apiNote use #addBar(Duration, ZonedDateTime, Num, Num, Num, Num, Num) to add bar data directly
      */
     void addBar(Bar bar);
 
+    /**
+     * Adds a bar at the end of the series.
+     * @param timePeriod the {@link Duration} of this bar
+     * @param endTime the {@link ZonedDateTime end time} of this bar
+     */
     void addBar(Duration timePeriod, ZonedDateTime endTime);
-    void addBar(ZonedDateTime endTime, Number openPrice, Number highPrice, Number lowPrice, Number closePrice);
-    void addBar(ZonedDateTime endTime, Number openPrice, Number highPrice, Number lowPrice, Number closePrice, Number volume);
-    void addBar(ZonedDateTime endTime, Number openPrice, Number highPrice, Number lowPrice, Number closePrice, Number volume, Number amount);
 
-    void addBar(ZonedDateTime endTime, String openPrice, String highPrice, String lowPrice, String closePrice);
-    void addBar(ZonedDateTime endTime, String openPrice, String highPrice, String lowPrice, String closePrice, String volume);
-    void addBar(ZonedDateTime endTime, String openPrice, String highPrice, String lowPrice, String closePrice, String volume, String amount);
+    default void addBar(ZonedDateTime endTime, Number openPrice, Number highPrice, Number lowPrice, Number closePrice){
+        this.addBar(endTime, numOf(openPrice), numOf(highPrice), numOf(lowPrice), numOf(closePrice), numOf(0), numOf(0));
+    }
 
-    void addBar(ZonedDateTime endTime, Num openPrice, Num highPrice, Num lowPrice, Num closePrice, Num volume);
-    void addBar(Duration timePeriod, ZonedDateTime endTime, Num openPrice, Num highPrice, Num lowPrice, Num closePrice, Num volume);
-    void addBar(Duration timePeriod, ZonedDateTime endTime, Num openPrice, Num highPrice, Num lowPrice, Num closePrice, Num volume, Num amount);
+    default void addBar(ZonedDateTime endTime, Number openPrice, Number highPrice, Number lowPrice, Number closePrice,
+                        Number volume){
+        this.addBar(endTime, numOf(openPrice), numOf(highPrice), numOf(lowPrice), numOf(closePrice), numOf(volume));
+    }
 
-    void addTrade(Number price, Number amount);
-    void addTrade(String price, String amount);
+    default void addBar(ZonedDateTime endTime, Number openPrice, Number highPrice, Number lowPrice, Number closePrice,
+                        Number volume, Number amount){
+        this.addBar(endTime, numOf(openPrice), numOf(highPrice), numOf(lowPrice), numOf(closePrice), numOf(volume),
+                numOf(amount));
+    }
+
+    default void addBar(ZonedDateTime endTime, String openPrice, String highPrice, String lowPrice, String closePrice){
+        this.addBar(endTime, numOf(new BigDecimal(openPrice)), numOf(new BigDecimal(highPrice)),
+                numOf(new BigDecimal(lowPrice)), numOf(new BigDecimal(closePrice)), numOf(0),
+                numOf(0));
+    }
+
+    default void addBar(ZonedDateTime endTime, String openPrice, String highPrice, String lowPrice, String closePrice,
+                        String volume){
+        this.addBar(endTime, numOf(new BigDecimal(openPrice)), numOf(new BigDecimal(highPrice)),
+                numOf(new BigDecimal(lowPrice)), numOf(new BigDecimal(closePrice)), numOf(new BigDecimal(volume)),
+                numOf(0));
+    }
+
+    default void addBar(ZonedDateTime endTime, String openPrice, String highPrice, String lowPrice, String closePrice,
+                        String volume, String amount){
+        this.addBar(endTime, numOf(new BigDecimal(openPrice)), numOf(new BigDecimal(highPrice)),
+                numOf(new BigDecimal(lowPrice)), numOf(new BigDecimal(closePrice)), numOf(new BigDecimal(volume)),
+                numOf(new BigDecimal(amount)));
+    }
+
+    default void addBar(ZonedDateTime endTime, Num openPrice, Num highPrice, Num lowPrice, Num closePrice, Num volume){
+        this.addBar(endTime, openPrice, highPrice, lowPrice, closePrice, volume, numOf(0));
+    }
+
+    /**
+     * Adds a new <code>Bar</code> to the time series.
+     * @param endTime end time of the bar
+     * @param openPrice the open price
+     * @param highPrice the high/max price
+     * @param lowPrice the low/min price
+     * @param closePrice the last/close price
+     * @param volume the volume (default zero)
+     * @param amount the amount (default zero)
+     */
+    void addBar(ZonedDateTime endTime, Num openPrice, Num highPrice, Num lowPrice, Num closePrice, Num volume,
+                Num amount);
+
+    /**
+     * Adds a new <code>Bar</code> to the time series.
+     * @param endTime end time of the bar
+     * @param openPrice the open price
+     * @param highPrice the high/max price
+     * @param lowPrice the low/min price
+     * @param closePrice the last/close price
+     * @param volume the volume (default zero)
+     */
+    void addBar(Duration timePeriod, ZonedDateTime endTime, Num openPrice, Num highPrice, Num lowPrice, Num closePrice,
+                Num volume);
+
+    /**
+     * Adds a new <code>Bar</code> to the time series.
+     * @param timePeriod the time period of the bar
+     * @param endTime end time of the bar
+     * @param openPrice the open price
+     * @param highPrice the high/max price
+     * @param lowPrice the low/min price
+     * @param closePrice the last/close price
+     * @param volume the volume (default zero)
+     * @param amount the amount (default zero)
+     */
+    void addBar(Duration timePeriod, ZonedDateTime endTime, Num openPrice, Num highPrice, Num lowPrice, Num closePrice,
+                Num volume, Num amount);
+
+    /**
+     * Adds a trade at the end of bar period.
+     * @param tradeVolume the traded volume
+     * @param tradePrice the price
+     */
+    default void addTrade(Number tradeVolume, Number tradePrice){
+        addTrade(numOf(tradeVolume), numOf(tradePrice));
+    }
+
+    /**
+     * Adds a trade at the end of bar period.
+     * @param tradeVolume the traded volume
+     * @param tradePrice the price
+     */
+    default void addTrade(String tradeVolume, String tradePrice){
+        addTrade(numOf(new BigDecimal(tradeVolume)), numOf(new BigDecimal(tradePrice)));
+    }
+
+    /**
+     * Adds a trade at the end of bar period.
+     * @param tradeVolume the traded volume
+     * @param tradePrice the price
+     */
     void addTrade(Num tradeVolume, Num tradePrice);
 
+    /**
+     * Adds a price to the last bar
+     * @param price the price for the bar
+     */
+    void addPrice(Num price);
+
+    default void addPrice(String price){
+        addPrice(new BigDecimal(price));
+    }
+
+    default void addPrice(Number price){
+        addPrice(numOf(price));
+    }
 
     /**
      * Returns a new TimeSeries implementation that is a subset of this TimeSeries implementation.