+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 ##(unreleased, `0.12-SNAPSHOT`)
 ### Breaking: 
    - `Decimal` class has been replaced by new `Num` interface. Enables using `Double`, `BigDecimal` and custom data types for calculations. 
-   - Big changes in `TimeSeries` and `BaseTimeSeries`. Multiple new `addBar(..)` functions in `TimeSeries` 
+   - Big changes in `TimeSeries` and `BaseTimeSeries`. Multiple new `addBar(..)` functions in `TimeSeries` allows to add data directly to the series
 
 ### Fixed
 ### Changed
@@ -12,18 +12,23 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **TimeSeries | Bar**: preferred way to add bar data to a `TimeSeries` is directly to the series via new `TimeSeries#addBar(time,open,high,..)` functions. It ensures to use the correct `Num` implementation of the series
 - **XlsTestsUtils**: now processes xls with one or more days between data rows (daily, weekly, monthly, etc).  Also handle xls #DIV/0! calculated cells (imported as NaN.NaN)
 - **CachedIndicator**: Last bar is not cached to support real time indicators
+- **TimeSeries | Bar**: added new `#addPrice(price)` function that adds price to (last) bar.
+- **Various Rules**: added constructor that provides `Number` parameters
+
 ### Added
 - **BaseTimeSeries.SeriesBuilder**: simplifies creation of BaseTimeSeries.
 - **Num**: Extracted interface of dropped `Decimal` class
 - **DoubleNum**: `Num` implementation to support calculations based on `double` primitive
 - **BigDecimalNum**: Default `Num` implementation of `BaseTimeSeries`
 - **TestUtils**: removed convenience methods for permuted parameters, fixed all unit tests
+- **TestUtils**: added parametrized abstract test classes to allow two test runs with `DoubleNum` and `BigDecimalNum`
+
 ### Removed/Deprecated
 - **Decimal**: _removed_. Replaced by `Num` interface
 - **TimeSeries#addBar(Bar bar)**: _deprecated_. Use `TimeSeries#addBar(Time, open, high, low, ...)`
 - **BaseTimeSeries**: _Consturctor_ `BaseTimeSeries(TimeSeries defaultSeries, int seriesBeginIndex, int seriesEndIndex)` _removed_. Use `TimeSeries.getSubseries(int i, int i)` instead
 
-##0.11 (released January 25, 2018)
+## 0.11 (released January 25, 2018)
 
 - **BREAKING**: Tick has been renamed to **Bar**